requirements.txt
langchain
langchain-openai
requests
bs4
openai
faiss-cpu
python-dotenv
beautifulsoup4
langchain_chroma
streamlit
langchain_community
chromadb==0.5.3
faiss-gpu
jq
langchain_aws
awscli
fastapi